Fix broken timers tests
These stopped working during the Vite upgrade so I skipped them to unblock myself.

When trying to fix them, I realized only one test actually relied on the fake clock (which is different from relying on fake timers - the logic looks at Date.now). So I moved the two other tests back into regular test files and then fixed that one test.
