+++ client/src/components/AuditAdmin/timers.test.tsx
@@ -6,7 +6,6 @@ import React, { ReactElement } from 'react'
 import { screen, act } from '@testing-library/react'
 import { Route } from 'react-router-dom'
 import FakeTimers from '@sinonjs/fake-timers'
-import userEvent from '@testing-library/user-event'
 import { QueryClientProvider } from 'react-query'
 import AuthDataProvider, { useAuthDataContext } from '../UserContext'
 import AuditAdminView from './AuditAdminView'
@@ -15,13 +14,7 @@ import {
   withMockFetch,
   createQueryClient,
 } from '../testUtilities'
-import {
-  aaApiCalls,
-  auditSettingsMocks,
-  roundMocks,
-  contestMocks,
-} from '../_mocks'
-import { sampleSizeMock } from './Setup/Review/_mocks'
+import { aaApiCalls, auditSettingsMocks, contestMocks } from '../_mocks'
 
 // AuditAdminView will only be rendered once the user is logged in, so
 // we simulate that.
@@ -42,27 +35,24 @@ const renderWithRoute = (route: string, component: ReactElement) =>
     }
   )
 
-const loadEach = [
-  aaApiCalls.getRounds([]),
-  aaApiCalls.getJurisdictions,
-  aaApiCalls.getContests(contestMocks.filledTargeted),
-  aaApiCalls.getSettings(auditSettingsMocks.all),
-]
-
-// TODO: Fix these tests after switching to Vite
-describe.skip('timers', () => {
+describe('timers', () => {
   let clock = FakeTimers.install()
   beforeEach(() => {
     clock = FakeTimers.install()
   })
-  afterEach(() => clock.uninstall())
+  afterEach(() => {
+    clock.uninstall()
+  })
   it('refreshes every five minutes on progress', async () => {
     const expectedCalls = [
       aaApiCalls.getUser,
-      ...loadEach,
-      ...loadEach,
+      aaApiCalls.getRounds([]),
+      aaApiCalls.getJurisdictions,
+      aaApiCalls.getContests(contestMocks.filledTargeted),
+      aaApiCalls.getSettings(auditSettingsMocks.all),
       aaApiCalls.getMapData,
-      ...loadEach,
+      aaApiCalls.getRounds([]),
+      aaApiCalls.getJurisdictions,
       aaApiCalls.getMapData,
     ]
     await withMockFetch(expectedCalls, async () => {
@@ -70,76 +60,15 @@ describe.skip('timers', () => {
       await act(async () => {
         await clock.nextAsync()
       })
-      screen.getByText('Will refresh in 5 minutes')
+      await screen.findByText('Will refresh in 5 minutes')
       await act(async () => {
         await clock.tickAsync(1000 * 60)
       })
-      screen.getByText('Will refresh in 4 minutes')
+      await screen.findByText('Will refresh in 4 minutes')
       await act(async () => {
-        // Five minutes minus the ten seconds we already ticked
-        await clock.tickAsync(1000 * (60 * 5 - 10))
+        await clock.tickAsync(1000 * 60 * 4)
       })
-      screen.getByText('Will refresh in 5 minutes')
-    })
-  })
-
-  it('shows a spinner while sample is being drawn', async () => {
-    const loadAfterLaunch = [
-      aaApiCalls.getRounds(roundMocks.drawSampleInProgress),
-      aaApiCalls.getJurisdictions,
-      aaApiCalls.getContests(contestMocks.filledTargeted),
-      aaApiCalls.getSettings(auditSettingsMocks.all),
-    ]
-    const expectedCalls = [
-      aaApiCalls.getUser,
-      ...loadAfterLaunch,
-      ...loadAfterLaunch,
-    ]
-    await withMockFetch(expectedCalls, async () => {
-      renderWithRoute('/election/1/progress', <AuditAdminViewWithAuth />)
-      await act(async () => {
-        await clock.nextAsync()
-      })
-      await screen.findByRole('heading', {
-        name: 'Drawing a random sample of ballots...',
-      })
-      screen.getByText(
-        'For large elections, this can take a couple of minutes.'
-      )
-    })
-  })
-
-  it('shows a spinner while sample sizes are computed', async () => {
-    const expectedCalls = [
-      aaApiCalls.getUser,
-      ...loadEach,
-      ...loadEach,
-      aaApiCalls.getSettings(auditSettingsMocks.all),
-      aaApiCalls.getJurisdictionFile,
-      ...loadEach,
-      aaApiCalls.getSettings(auditSettingsMocks.all),
-      aaApiCalls.getJurisdictions,
-      aaApiCalls.getJurisdictionFile,
-      aaApiCalls.getContests(contestMocks.filledTargeted),
-      aaApiCalls.getSampleSizes(sampleSizeMock.calculating),
-      aaApiCalls.getSampleSizes(sampleSizeMock.ballotPolling),
-    ]
-    await withMockFetch(expectedCalls, async () => {
-      renderWithRoute('/election/1/setup', <AuditAdminViewWithAuth />)
-      await act(async () => {
-        await clock.nextAsync()
-      })
-      await screen.findByRole('heading', { name: 'Audit Setup' })
-      userEvent.click(screen.getByText('Review & Launch'))
-      await screen.findByRole('heading', { name: 'Sample Size' })
-      screen.getByText('Loading sample size options...')
-
-      await act(async () => {
-        await clock.tickAsync(1000 * 60)
-      })
-      screen.getByText(
-        'Choose the initial sample size for each contest you would like to use for Round 1 of the audit from the options below.'
-      )
+      await screen.findByText('Will refresh in 5 minutes')
     })
   })
 })
